#15
Addition of Sales.hpp and Sales.cpp.
These have the classes for Sales and SalesList. Sales stores the data of a sale and SalesList Holds all of the sales in a vector, as well as being able to add new sales, creating new sales .cvs files, loading .cvs siles, or saving .cvs files.