#15
Added Sales header and source files. The Sales Class just holds all of the data relevant for a given sale. The SalesList class hold a vector of sales to keep track of all the sales while the program is running. The vector can be loaded with information from a file, can create a new file file, or save the sales to a file.